Add experimental Custom Area import flow
