--- .beta30-trigger
@@ -1,1 +1,0 @@
-beta30 custom area builder trigger
